BACKEND/requirements.txt
argon2-cffi==23.1.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Cors==5.0.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.1.1
marshmallow-sqlalchemy==1.4.2
mysql-connector-python==9.5.0
python-dotenv==1.0.1
python-jose==3.3.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
